+++ src/experiments/single.rs
@@ -20,11 +20,11 @@ use rand::SeedableRng;
 
 /// Generate the group layout to use for this experiment.
 fn group_layout(
-    num_cells: usize,
+    num_cells: u32,
     char_quants: &CharQuantities,
     bottom_left: (Length, Length),
-    width: usize,
-    height: usize,
+    width: u32,
+    height: u32,
 ) -> Result<GroupBBox, String> {
     // specify initial location of group bottom left
     let bottom_left = V2D {
@@ -61,13 +61,13 @@ fn cell_groups(
 }
 
 /// Generate CAL values between different cells.
-fn gen_cal_mat() -> SymCcDat<f64> {
-    SymCcDat::<f64>::new(2, 60.0)
+fn gen_cal_mat() -> SymCcDat<f32> {
+    SymCcDat::<f32>::new(2, 60.0)
 }
 
 /// Generate CIL values between different cells.
-fn gen_cil_mat() -> SymCcDat<f64> {
-    SymCcDat::<f64>::new(2, 60.0)
+fn gen_cil_mat() -> SymCcDat<f32> {
+    SymCcDat::<f32>::new(2, 60.0)
 }
 
 /// Generate raw world parameters, in particular, how
